+++ src/com/PluginBase/Main.java
@@ -4,7 +4,12 @@ import org.bukkit.plugin.java.JavaPlugin;
 
 public class Main extends JavaPlugin {
 	
-	public static final String version = "1.0";
+	/*
+	 * TODO:
+	 * - stop using an invisible prefix for the clickable inventory and switch to just saving the object
+	 */
+	
+	public static final String version = "1.2";
 	
 	@Override
 	public void onEnable() {