+++ changelog.md
@@ -1,7 +1,7 @@
 # Version 1.1
 
 - Added SoundEmitter.java
-- Added ``emitParticles(..., DustOptions dustOptions, ...)``
+- Added ``emitParticles(..., DustOptions dustOptions, ...)`` and ``emitParticles(..., BlockData blockData, ...)``
 
 # Version 1.0