+++ changelog.md
@@ -1,7 +1,8 @@
 # Version 1.1
 
 - Added SoundEmitter.java
-- Added ``emitParticles(..., DustOptions dustOptions, ...)`` and ``emitParticles(..., BlockData blockData, ...)`` to add support for particles ``Particle.REDSTONE``, ``Particle.ITEM_CRACK``, ``Particle.BLOCK_CRACK`` and ``Particle.BLOCK_DUST``
+- Added broadcast to administrators when an error message is sent to the console
+- Added ``emitParticles(..., DustOptions dustOptions, ...)`` and ``emitParticles(..., BlockData blockData, ...)`` to add support for particles ``Particle.REDSTONE``, ``Particle.ITEM_CRACK``, ``Particle.BLOCK_CRACK`` and ``Particle.BLOCK_DUST``. There is also a new error message when trying to spawn these particles without the necessary information. NOTE: This does not yet work for continuous emission
 
 # Version 1.0